requirements.txt

streamlit
pillow
qrcode[pil]
reportlab
pandas
